backend/requirements.txt
flask
gunicorn
python-dotenv
psycopg2-binary
bcrypt
pyjwt
pytest
python-dotenv